normalises moments to fix numerical problems
sometimes the circle fits were poor and this is attributed to numerical
issues in eigenvalue decomposition when the moment based matrix
elements were several orders of magnitude smaller than n (the length of
the data).

by normalising and switching to an alternative B matrix based on these
same orders of magntiude, circle fits seem to be more reliable. I am
not sure what exact difference is made by the change of matrix but this
update seems to work more reliably on real resonance data.